live_location_tracking/requirement.txt
Django>=4.1,<5
djangorestframework>=3.12
channels>=3.0
channels_redis>=4.0
djangorestframework_simplejwt>=4.0
